split the custom type module into multiple files
